# project/api/server/api.yaml
openapi: '3.1.0'
info:
  version: 1.0.0
  title: Admin API
  description: The admin-only portion of the API, which has its own separate OpenAPI spec
paths:
  /:
    get:
      summary: Index
      operationId: getIndex
      responses:
        200:
          $ref: '#/components/responses/htmx'
  /login/{provider}:
    get:
      summary: Process a user login with given provider
      operationId: getLogin
      tags:
        - login
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      responses:
        307:
          description: Redirect to provider
        404:
          description: No provider to satisfy login
        500:
          description: Problem with request

  /login/{provider}/callback:
    get:
      summary: Callback for login provider
      operationId: getLoginCallback
      tags:
        - login
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        307:
          description: Redirect to home
        401:
          description: Unauthorized
        404:
          description: No provider to satisfy login
        500:
          description: Problem with request

  /logout/{provider}:
    get:
      summary: Logout handler for provider
      operationId: getLogout
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      responses:
        307:
          description: Redirect to provider
        500:
          description: Problem with request

  /signup:
    get:
      summary: User sign up
      operationId: getSignup
      responses:
        200:
          description: Show sign up page
          content:
            text/html: {}
        400:
          description: Bad request (not htmx)
        500:
          description: Error rendering signup form
    post:
      summary: Process sign up request
      operationId: postSignup
      requestBody:
        description: New user information
        content:
          multipart/form-data:
            schema:
              $ref: '../models/api.yaml#/components/schemas/APIUser'
      responses:
        200:
          $ref: '#/components/responses/htmx'
        500:
          description: Error processing signup

  /signup/validate:
    post:
      summary: Validate sign up
      operationId: postSignupValidate
      responses:
        200:
          description: Validation completed
          content:
            text/html: {}
      requestBody:
        description: Item to validate
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '../models/api.yaml#/components/schemas/APIUser'

  /home:
    get:
      summary: Display user home page
      operationId: getHome
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with request
  /search:
    post:
      summary: Issue a user search request
      operationId: Search
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with request
      # requestBody:
      #   description: filters for results
      #   content:
      #     application/x-www-form-urlencoded:
      #       schema:
      #         $ref: '../models/api.yaml#/components/schemas/Filters'
  /home/settings:
    get:
      summary: Show user settings modal
      operationId: getHomeSettings
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with request
  /home/subscription/add:
    get:
      summary: Add a new subscription
      operationId: getSubscriptionAdd
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with request
    post:
      summary: Process a new subscription
      operationId: postSubscriptionAdd
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with subscription
  /home/subscription/validate:
    post:
      summary: Validate a subscription
      operationId: postSubscriptionValidate
      responses:
        200:
          $ref: '#/components/responses/htmx'
      requestBody:
        description: Subscription to validate
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '../models/api.yaml#/components/schemas/APISubscription'
  /home/subscription/edit/{subID}:
    get:
      summary: Edit a new subscription
      operationId: getSubscriptionEdit
      parameters:
        - name: subID
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with request
    post:
      summary: Process a subscription edit
      operationId: postSubscriptionEdit
      parameters:
        - name: subID
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/htmx'
        401:
          description: Unauthorized
        500:
          description: Problem with request
  /home/feeds:
    get:
      summary: list feeds filtered by given parameters
      operationId: getFeedList
      parameters:
        - $ref: '#/components/parameters/FeedsCookie'
      responses:
        200:
          $ref: '#/components/responses/htmx'
    post:
      summary: update feeds list filters
      operationId: updateFeedList
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '../models/api.yaml#/components/schemas/FeedsFilters'
      responses:
        200:
          $ref: '#/components/responses/htmx'
  /home/items:
    get:
      summary: list items with filtered by given parameters
      operationId: listItems
      parameters:
        - $ref: '#/components/parameters/ItemsCookie'
          in: cookie
      responses:
        200:
          $ref: '#/components/responses/htmx'
    post:
      summary: update items list filters
      operationId: updateItemsList
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '../models/api.yaml#/components/schemas/ItemsFilters'
  /home/{feed}:
    get:
      summary: display a feed
      operationId: showFeed
      parameters:
        - $ref: '../models/api.yaml#/components/parameters/FeedID'
          required: true
      responses:
        200:
          $ref: '#/components/responses/htmx'
  /home/{feed}/{item}:
    get:
      summary: display an item
      operationId: showItem
      parameters:
        - $ref: '../models/api.yaml#/components/parameters/FeedID'
          required: true
        - $ref: '../models/api.yaml#/components/parameters/ItemID'
          required: true
      responses:
        200:
          $ref: '#/components/responses/htmx'
components:
  parameters:
    FeedID:
      name: feed
      in: path
      schema:
        type: string
    ItemID:
      name: item
      in: path
      schema:
        type: string
    FeedsCookie:
      description: contains settings for the feeds list pane.
      name: feedsFilters
      in: cookie
      schema:
        $ref: '../models/api.yaml#/components/schemas/FeedsFilters'
    ItemsCookie:
      description: contains settings for the items list pane.
      name: itemsFilters
      in: cookie
      schema:
        $ref: '../models/api.yaml#/components/schemas/ItemsFilters'
  responses:
    htmx:
      description: standard htmx powered response.
      content:
        'text/html;charset=utf-8': {}